Adicionado funcionalidade de editar um cliente, Melhorias de código, correção no parse de data

+++ src/main/java/com/loja/suplementos/customer/CustomerService.java
@@ -47,11 +47,29 @@ public class CustomerService {
         return customerRepository.findById(id).orElse(null);
     }
 
-    public void update(Customer customer) {
-        customerRepository.update(customer);
+    public void update(Map<String, String> params) {
+        var email = params.get("email");
+        var cpf = params.get("cpf");
+
+        Customer existingCustomer = customerRepository
+            .findById(Long.valueOf(params.get("id")))
+            .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado"));
+
+        if (!existingCustomer.getCpf().equals(cpf)) throw new IllegalArgumentException("Não é possível alterar o CPF");
+        if (!existingCustomer.getEmail().equals(email)) throw new IllegalArgumentException("Não é possível alterar o email");
+
+        existingCustomer.setName(params.get("name"));
+        existingCustomer.setLastName(params.get("lastName"));
+        existingCustomer.setBirthDate(Utils.convertStringToDate(params.get("birthDate")));
+        existingCustomer.setPhone(params.get("phoneNumber"));
+
+        customerRepository.update(existingCustomer);
     }
 
-    public void delete(Customer customer) {
+    public void delete(Long customerId) {
+        Customer customer = this.findById(customerId);
+        if (customer == null) throw new IllegalArgumentException("Cliente não encontrado");
+
         customerRepository.delete(customer);
     }
